+++ lib/led_auto_mode/led_auto_mode.h
@@ -74,7 +74,7 @@ class Pixels {
         float pulseIndex = 0;
 
         // frequency divider
-        uint8_t freqdiv = 1;
+        float freqdiv = 1;
 
     private:
